requirements.txt
# Core dependencies
streamlit==1.42.2
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1

# Gemini and LangGraph
google-generativeai==0.8.4
langgraph==0.2.74
langchain-core==0.3.40
langchain==0.3.19

# Utilities
tqdm==4.67.1
pydantic==2.10.6
requests==2.32.3
